+++ components/Box/utils/getShallotProp.ts
@@ -12,6 +12,8 @@ import {
 } from '@shallot-ui/core'
 
 import { BoxProps, BoxShallot } from '../types'
+import { useTheme } from 'styled-components'
+import { Variant } from '@shallot-ui/theme'
 
 const getShallotProp = <T>(props: BoxProps<T>): BoxShallot => {
   const baseShallot: BoxShallot = applyStyles({
@@ -20,7 +22,14 @@ const getShallotProp = <T>(props: BoxProps<T>): BoxShallot => {
     ...props.shallot,
   })
 
+  const theme = useTheme()
+  const { variant = 'default' } = props
+  const themeVariant = theme?.variants?.Box?.[variant] as
+    | Variant<BoxShallot>
+    | undefined
+
   return applyStyles(baseShallot, {
+    // ...theme?.variants?.Box?.[props.variant],
     ...getAlignmentShallot(baseShallot.flexDirection, props),
     ...getBorderShallot(props),
     ...getBackgroundColorShallot(props),
@@ -30,6 +39,9 @@ const getShallotProp = <T>(props: BoxProps<T>): BoxShallot => {
     ...getMarginShallot(props),
     ...getRadiusShallot(props),
     ...getSizingShallot(props),
+
+    // Variants (overrides)
+    ...themeVariant,
   })
 }