+++ components/Column/utils/getShallotProp.ts
@@ -12,6 +12,8 @@ import {
 } from '@shallot-ui/core'
 
 import { ColumnProps, ColumnShallot } from '../types'
+import { useTheme } from 'styled-components'
+import { Variant } from '@shallot-ui/theme'
 
 const getShallotProp = <T>(props: ColumnProps<T>): ColumnShallot => {
   const baseShallot: ColumnShallot = applyStyles({
@@ -19,6 +21,12 @@ const getShallotProp = <T>(props: ColumnProps<T>): ColumnShallot => {
     flexDirection: 'column',
     ...props.shallot,
   })
+
+  const theme = useTheme()
+  const { variant = 'default' } = props
+  const themeVariant = theme?.variants?.Box?.[variant] as
+    | Variant<ColumnShallot>
+    | undefined
 
   return applyStyles(baseShallot, {
     ...getAlignmentShallot(baseShallot.flexDirection, props),
@@ -30,6 +38,9 @@ const getShallotProp = <T>(props: ColumnProps<T>): ColumnShallot => {
     ...getMarginShallot(props),
     ...getRadiusShallot(props),
     ...getSizingShallot(props),
+
+    // Variants (overrides)
+    ...themeVariant,
   })
 }