+++ components/Fold/index.tsx
@@ -1,8 +1,7 @@
 import { ComponentType } from 'react'
 import { DefaultTheme } from 'styled-components'
-import { AllColorShades } from '@shallot-ui/theme'
+import { AllColorShades, ShallotProp } from '@shallot-ui/theme'
 import {
-  ShallotProp,
   applyStyles,
   getColorShade,
   getFullHeight,