+++ packages/theme/src/theme/variants/index.ts
@@ -1,0 +1,5 @@
+import { Variant } from '../../types'
+
+export const DEFAULT_VARIANTS = {} as {
+  [component: string]: { [variant: string]: Variant }
+}